test/iscsi: Add SPDK iSCSI target performance test with VPP enabled
To run this test 3 IP addresses must be speciefied:
--ip_tgt_int - address of target interface used for iSCSI
--ip_init_int - adress of initiator interface used for iSCSI
--ip_init - address to initiator

+++ test/iscsi_tgt/perf/iscsi_target.sh
@@ -12,6 +12,7 @@ FIO_PATH="/usr/src/fio"
 DISKNO="ALL"
 CPUMASK=0x02
 NUM_JOBS=1
+USE_VPP=false
 . $(readlink -e "$(dirname $0)/../common.sh")
 
 # Performance test for iscsi_tgt, run on devices with proper hardware support (target and inititator)
@@ -29,10 +30,12 @@ function usage()
 	echo "    --fiopath=PATH      Path to fio directory on initiator. [default=$FIO_PATH]"
 	echo "    --disk_no=INT,ALL   Number of disks to test on, if =ALL then test on all found disk. [default=$DISKNO]"
 	echo "    --cpumask=HEX       The parameter given is a bit mask of allowed CPUs the job may run on. [default=$CPUMASK]"
-	echo "    --numjobs=INT     Create the specified number of clones of this job. Each clone of job is spawned as an independent thread or process. [default=$NUM_JOBS]"
+	echo "    --numjobs=INT       Create the specified number of clones of this job. Each clone of job is spawned as an independent thread or process. [default=$NUM_JOBS]"
 	echo "    --target_ip=IP      The IP adress of target used for test."
 	echo "    --initiator_ip=IP   The IP adress of initiator used for test."
 	echo "    --init_mgmnt_ip=IP  The IP adress of initiator used for communication."
+	echo "    --with-vpp          Test iscsi_tgt with vpp enabled."
+	echo "    --dpdk_drv=STR      If vpp is enabled: name of the DPDK driver to bind to target NIC."
 }
 
 while getopts 'h-:' optchar; do
@@ -53,6 +56,8 @@ while getopts 'h-:' optchar; do
 			target_ip=*) TARGET_IP="${OPTARG#*=}" ;;
 			initiator_ip=*) INITIATOR_IP="${OPTARG#*=}" ;;
 			init_mgmnt_ip=*) IP_I_SSH="${OPTARG#*=}" ;;
+			with-vpp) USE_VPP=true ;;
+			dpdk_drv=*) DPDK_DRV="${OPTARG#*=}" ;;
 			*) usage $0 echo "Invalid argument '$OPTARG'"; exit 1 ;;
 		esac
 		;;
@@ -64,6 +69,11 @@ done
 . $(readlink -e "$(dirname $0)/../../common/autotest_common.sh") || exit 1
 testdir=$(readlink -f $(dirname $0))
 rootdir=$(readlink -f $testdir/../../..)
+trap "print_backtrace; exit 1" ERR SIGTERM SIGABRT
+TGT_NETMASK=$(ip a | grep $TARGET_IP | awk '{print $2}')
+TGT_INT=$(ip a | grep $TARGET_IP | awk '{print $NF}')
+TGT_INT_ADDR=$($rootdir/dpdk/usertools/dpdk-devbind.py --status | grep $TGT_INT | awk '{print $1}')
+TGT_INT_DRV=$($rootdir/dpdk/usertools/dpdk-devbind.py --status | grep -oP "(?<=$TGT_INT drv=).*" | awk '{print $1}')
 
 if [ -z "$TARGET_IP" ]; then
 	error "No IP adress of iscsi target is given"
@@ -77,6 +87,10 @@ if [ -z "$IP_I_SSH" ]; then
 	error "No IP adress of initiator is given"
 fi
 
+if $USE_VPP && [ -z "$DPDK_DRV" ]; then
+	error "No name of DPDK driver is given"
+fi
+
 if [[ $EUID -ne 0 ]]; then
 	error "INFO: Go away user come back as root"
 fi
@@ -85,18 +99,57 @@ function ssh_initiator(){
 	ssh -i $HOME/.ssh/spdk_vhost_id_rsa root@$IP_I_SSH "$@"
 }
 
+function setup_vpp_iscsi()
+{
+	if ! lsmod | grep $DPDK_DRV; then
+		modprobe $DPDK_DRV
+	fi
+
+	ip link set dev $TGT_INT down
+	sleep 1
+	$rootdir/dpdk/usertools/dpdk-devbind.py --bind=$DPDK_DRV $TGT_INT
+	sleep 1
+	systemctl start vpp
+	sleep 7
+	vpp_int_name=$(vppctl show int | grep Ethernet | awk '{print $1}')
+	vppctl set int ip address $vpp_int_name $TGT_NETMASK
+	vppctl set int state $vpp_int_name up
+	sleep 1
+}
+
+function clean_vpp_iscsi()
+{
+	sleep 1
+	vppctl set int state $vpp_int_name down
+	systemctl stop vpp
+	sleep 5
+	$rootdir/dpdk/usertools/dpdk-devbind.py --bind=$TGT_INT_DRV $TGT_INT_ADDR
+	sleep 1
+	ip link set dev $TGT_INT up
+}
+
 NETMASK=$INITIATOR_IP/32
 rpc_py="python $rootdir/scripts/rpc.py -s $testdir/rpc_iscsi.sock"
 iscsi_fio_results="$testdir/perf_output/iscsi_fio.json"
-rm -rf $iscsi_fio_results
-mkdir -p $testdir/perf_output
-touch $iscsi_fio_results
 
 timing_enter run_iscsi_app
-$rootdir/app/iscsi_tgt/iscsi_tgt -r $testdir/rpc_iscsi.sock -c $testdir/iscsi.conf &
+
+if $USE_VPP; then
+	iscsi_fio_results="$testdir/perf_output/iscsi_vpp_fio.json"
+	trap "clean_vpp_iscsi; print_backtrace; exit 1" ERR SIGTERM SIGABRT
+	setup_vpp_iscsi
+fi
+
+$rootdir/app/iscsi_tgt/iscsi_tgt -r $testdir/rpc_iscsi.sock -c $testdir/iscsi.conf -m 0x2 &
 pid=$!
 waitforlisten "$pid" "$testdir/rpc_iscsi.sock"
-trap "rm -f $testdir/perf.job; killprocess $pid; print_backtrace; exit 1" ERR SIGTERM SIGABRT
+
+if $USE_VPP; then
+	trap "killprocess $pid; clean_vpp_iscsi; print_backtrace; exit 1" ERR SIGTERM SIGABRT
+else
+	trap "rm -f $testdir/perf.job; killprocess $pid; print_backtrace; exit 1" ERR SIGTERM SIGABRT
+fi
+
 sleep 1
 timing_exit run_iscsi_app
 
@@ -131,10 +184,19 @@ rm -f $testdir/perf.job
 timing_exit iscsi_config
 
 timing_enter iscsi_initiator
-ssh_initiator bash -s - < $testdir/iscsi_initiator.sh $FIO_PATH $TARGET_IP
+ssh_initiator bash -s - < $testdir/iscsi_initiator.sh $FIO_PATH $TARGET_IP $USE_VPP
 timing_exit iscsi_initiator
 
-ssh_initiator "cat perf_output/iscsi_fio.json" | cat  > $iscsi_fio_results
-ssh_initiator "rm -rf perf_output perf.job"
-
 killprocess $pid
+
+rm -rf $iscsi_fio_results
+mkdir -p $testdir/perf_output
+touch $iscsi_fio_results
+if $USE_VPP; then
+	clean_vpp_iscsi
+	ssh_initiator "cat perf_output/iscsi_vpp_fio.json" | cat  > $iscsi_fio_results
+else
+	ssh_initiator "cat perf_output/iscsi_fio.json" | cat  > $iscsi_fio_results
+fi
+
+ssh_initiator "rm -rf perf_output perf.job"
